Code/src/p300_service/requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
async-timeout==3.0.0
asyncio==3.4.3
attrs==18.1.0
bitstring==3.1.5
certifi==2018.8.24
chardet==3.0.4
click==6.7
cycler==0.10.0
dnspython==1.15.0
enum34==1.1.6
eventlet==0.24.1
Flask==1.0.2
greenlet==0.4.14
httptools==0.0.11
idna==2.7
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.5
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mne==0.16.2
monotonic==1.5
multidict==4.3.1
muselsl==1.0.2
numpy==1.15.1
pandas==0.23.4
pexpect==4.6.0
ptyprocess==0.6.0
pygatt==3.1.1
pylsl==1.12.2
pyparsing==2.2.0
pyriemann==0.2.5
pyserial==3.4
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
requests==2.19.1
sanic==0.8.3
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
socketIO-client==0.7.2
urllib3==1.23
websocket-client==0.53.0
websockets==5.0.1
Werkzeug==0.14.1
yarl==1.2.6
